[⚙️ chore] 메뉴 관리 관련 컴포넌트 스타일 파일 삭제 #1
