+++ src/components/MenuManagement/Header.tsx
@@ -1,6 +1,24 @@
+import React from 'react';
 import styled from 'styled-components';
+function Header() {
+	return (
+		<HeadWrapper>
+			<ImageContainer>
+				<img src="/assets/admin/back_light.svg" alt="뒤로가기" />
+			</ImageContainer>
+			<h1>메뉴관리</h1>
+			<ButtonContainer>
+				<button type="button">카레고리 관리</button>
+				<button type="button">메뉴 추가</button>
+			</ButtonContainer>
+		</HeadWrapper>
+	);
+}
 
-export const HeadWrapper = styled.header`
+export default Header;
+
+const HeadWrapper = styled.header`
+	background-color: ${({ theme }) => theme.textColor.white};
 	height: 80px;
 	display: flex;
 	align-items: center;
@@ -12,12 +30,12 @@ export const HeadWrapper = styled.header`
 	}
 `;
 
-export const ImageContainer = styled.button`
+const ImageContainer = styled.button`
 	margin-left: 33px;
 	width: 60px;
 `;
 
-export const ButtonContainer = styled.div`
+const ButtonContainer = styled.div`
 	button {
 		width: 140px;
 		height: 56px;
