+++ src/components/MenuManagement/MenuList.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import styled from 'styled-components';
-import MenuItem from '../../MenuItem';
+import MenuItem from '../MenuItem';
 
 function MenuList() {
 	return (
